Store mongod host / port in the configuration.

+++ spec/spec_helper.rb
@@ -1,6 +1,8 @@
 $:.unshift(File.dirname(__FILE__))
 $:.unshift(File.join(File.dirname(__FILE__), "..", "lib"))
 
+require "mongo_browser"
+
 require "rspec"
 require "capybara"
 require "capybara/rspec"
@@ -16,23 +18,23 @@ ensure
   server.close if server
 end
 
-ENV["MONGODB_PORT"] = find_available_port.to_s
+MongoBrowser.mongodb_host = "localhost"
+MongoBrowser.mongodb_port = find_available_port.to_s
 ENV["MONGODB_DBPATH"] = "/tmp/mongo_browser/db"
 
-require "mongo_browser"
 Capybara.app = MongoBrowser::Application
 
 def start_mongodb_server
-  if `lsof -i :#{ENV["MONGODB_PORT"]}`.size == 0
+  if `lsof -i :#{MongoBrowser.mongodb_port}`.size == 0
     clean_up
     FileUtils.mkdir_p(ENV["MONGODB_DBPATH"])
 
-    `mongod --port #{ENV["MONGODB_PORT"]} --dbpath #{ENV["MONGODB_DBPATH"]} --fork --logpath #{ENV["MONGODB_DBPATH"]}/../db.log`
+    `mongod --port #{MongoBrowser.mongodb_port} --dbpath #{ENV["MONGODB_DBPATH"]} --fork --logpath #{ENV["MONGODB_DBPATH"]}/../db.log`
   end
 end
 
 def load_fixtures
-  connection = Mongo::Connection.new("localhost", ENV["MONGODB_PORT"])
+  connection = Mongo::Connection.new(MongoBrowser.mongodb_host, MongoBrowser.mongodb_port)
 
   # Delete all databases
   connection.database_names do |db_name|
@@ -47,8 +49,8 @@ end
 def clean_up
   out = `ps aux | grep "mongod"`
   out.lines.each do |line|
-    if line.include?("--dbpath #{ENV["MONGODB_DBPATH"]}")
-      pid = line.split(/\s/)[2]
+    if line.include?("mongod") and line.include?("--dbpath #{ENV["MONGODB_DBPATH"]}")
+      pid = line.split(/\s/)[1]
       `kill -9 #{pid}`
     end
   end
